requirements.txt
streamlit==0.84.2
numpy
pandas
httpx-oauth==0.3.5
google-auth-oauthlib==0.4.1
google-api-python-client
google-auth
asyncio
pyjwt
environ
markdown
matplotlib
plotly
python-dotenv
scipy
markdown
folium
pydeck
streamlit_folium
requests
rq
redis